Move dataset_subscription_task to dataset sharing
