+++ backend/dataall/modules/dataset_sharing/tasks/dataset_subscription_task.py
@@ -15,7 +15,7 @@ from dataall.modules.dataset_sharing.services.share_notification_service import 
 from dataall.modules.datasets.aws.sns_dataset_client import SnsDatasetClient
 from dataall.modules.datasets.db.dataset_location_repositories import DatasetLocationRepository
 from dataall.modules.datasets.db.dataset_table_repositories import DatasetTableRepository
-from dataall.modules.datasets.tasks.subscriptions import poll_queues
+from dataall.modules.dataset_sharing.tasks.subscriptions import poll_queues
 from dataall.modules.datasets_base.db.dataset_repositories import DatasetRepository
 from dataall.modules.datasets_base.db.dataset_models import DatasetStorageLocation, DatasetTable, Dataset
